style: 해시태그 디데일 뷰 작업
