+++ lib/constants/color_constants.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class AppColors {
   /// app color
   static const Color primaryColor = Color(0xffFD6846);
-  static const Color secondaryColor = Color(0xffFFF4EE);
+  static const Color moaSecondary = Color(0xffFFF4EE);
   static const Color backgroundColor = Color(0xffFFE296);
   static const Color folderColorFAE3CB = Color(0xffFAE3CB);
   static const Color folderColorFFD4D7 = Color(0xffFFD4D7);
@@ -17,7 +17,7 @@ class AppColors {
 
   ///font color
   static const Color fontTitleColor = Color(0xff000000);
-  static const Color fontDescriptionColor = Color(0xff828282);
+  static const Color moaDescription = Color(0xff828282);
   static Color moaOpacity30 = const Color(0xff000000).withOpacity(0.3);
 
   /// deprecated
